[win32] storage: move implementation to platform file
