+++ xbmc/platform/win32/storage/CMakeLists.txt
@@ -2,4 +2,4 @@ set(SOURCES Win32StorageProvider.cpp)
 
 set(HEADERS Win32StorageProvider.h)
 
-core_add_library(storage_windows)
+core_add_library(platform_win32_storage)
