+++ xbmc/platform/win32/storage/Win32StorageProvider.cpp
